backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.112.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
motor==3.5.1
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.38.2
tomli==2.0.1
trio==0.26.2
typing_extensions==4.12.2
uvicorn==0.30.6
